Update README.md
Project Summary:
We are going to attempt to program a choose your own adventure game. It will combine all the major fandoms: Lord of the Rings, Harry Potter, Pokemon, and Hunger Games. The user will be able to travel to the different fandom realms and make decisions to finish the quest. They'll be able to earn certain items to help them defeat the final boss. The user will traverse through a series of arrays, each containing a different world.  There will be a Charater class to customize the character as well as the functions that will make the journey possible. 

Dependencies:
An advanced C++ compiler (c++11) to handle the stoi function
For the potions riddle in the harry potter quest, here's a picture to make it a little easier to visualize:
http://img1.wikia.nocookie.net/__cb20110906232228/harrypotter/images/e/e5/Potions_Riddle.png

System Requirements: 
No additonal requirements needed.

Group Members:
Sonia Szeton
Connor Fitzgerald

Contributors:
N/A

Open issues/Bugs:
Don't know how to implement chess game for harry potter chamber. 
Can't link the header file to the driver file even when #include "Character.h" is used
It would be cool to have the picture for the potions riddle (link above) pop up in a new window when the player gets to that chamber